add axis labels to disaggregationBarChart; sizing changes; expose numTicks for axes;
