+++ src/DisaggregationBarChart/DisaggregationBarChart.types.ts
@@ -3,3 +3,16 @@ export interface BarData {
   y1: number[];
   y2: number[];
 }
+
+export type DisaggregationBarChartProps = {
+  barData: BarData;
+  colourArray: string[];
+  width: number;
+  height: number;
+  events?: boolean;
+  verticalMargin: number;
+  xLabel: string;
+  yLabel: string;
+  xNumTicks?: number;
+  yNumTicks?: number;
+};
